Restaurant application
The Site displays the food menu available in the restaurant and recipe
of the food item and the customer feedback is also displayed for each
food item. Customers can place the order of food menu online and book a
table before they visit the restaurant else they can also order the
food to their place from the Site.